Refactor placeholders expansion
The replacement function is moved to utils/utils.js.
It is used whereever placeholders are used.
