+++ src/components/locator.ts
@@ -3,6 +3,7 @@ import * as fs from 'fs'
 import * as path from 'path'
 import * as cp from 'child_process'
 import * as synctexjs from './synctex'
+import {replaceArgumentPlaceholders} from '../utils/utils'
 
 import {Extension} from '../main'
 import {ClientRequest} from '../../viewer/components/protocol'
@@ -431,16 +432,13 @@ export class Locator {
         const configuration = vscode.workspace.getConfiguration('latex-workshop')
         const command = configuration.get('view.pdf.external.synctex.command') as string
         let args = configuration.get('view.pdf.external.synctex.args') as string[]
-        const rootFileParsed = path.parse(rootFile)
-        const docfile = rootFileParsed.name
-        const dir = path.normalize(rootFileParsed.dir).split(path.sep).join('/')
-        const doc = path.join(dir, docfile)
         if (args) {
-            args = args.map(arg => arg.replace(/%DOC%/g, doc)
-                                      .replace(/%DOCFILE%/g, docfile)
-                                      .replace(/%PDF%/g, pdfFile)
-                                      .replace(/%LINE%/g, line.toString())
-                                      .replace(/%TEX%/g, texFile))
+            args = args.map(arg => {
+                return replaceArgumentPlaceholders(rootFile, this.extension.builder.tmpDir)(arg)
+                        .replace(/%PDF%/g, pdfFile)
+                        .replace(/%LINE%/g, line.toString())
+                        .replace(/%TEX%/g, texFile)
+            })
         }
         this.extension.manager.setEnvVar()
         cp.spawn(command, args)
