+++ src/components/manager.ts
@@ -80,19 +80,10 @@ export class Manager {
             return './'
         }
 
-        const texPathParsed = path.parse(texPath)
-        const docfile = texPathParsed.name
-        const dir = path.normalize(texPathParsed.dir).split(path.sep).join('/')
-        const doc = path.join(dir, docfile)
         const configuration = vscode.workspace.getConfiguration('latex-workshop')
-        const docker = configuration.get('docker.enabled')
         const outDir = configuration.get('latex.outDir') as string
-        const out = outDir.replace(/%DOC%/g, docker ? docfile : doc)
-                  .replace(/%DOCFILE%/g, docfile)
-                  .replace(/%DIR%/g, docker ? './' : dir)
-                  .replace(/%TMPDIR%/g, this.extension.builder.tmpDir)
+        const out = utils.replaceArgumentPlaceholders(texPath, this.extension.builder.tmpDir)(outDir)
         return path.normalize(out).split(path.sep).join('/')
-
     }
 
     get rootDir() {
